Call upon exterior matchups.js file within routes
