+++ views/index.pug
@@ -23,7 +23,7 @@ html
 
 		ul Wesley Chipeco
 			li 
-				a(href= '/owner2_matchups=2015_2016') Trifecta Season 1: 2015-2016
+				a(href= '/owner/2/matchups/2015/2016') Trifecta Season 1: 2015-2016
 
 			li 
-				a(href= '/owner2_matchups=2016_2017') Trifecta Season 2: 2016-2017
+				a(href= '/owner/2/matchups/2016/2017') Trifecta Season 2: 2016-2017
